+++ src/RemixDevTools/index.ts
@@ -1,2 +1,6 @@
 export { RemixDevTools } from "./RemixDevTools";
 export { useRemixForgeSocketExternal as useRemixForgeSocket } from "./hooks/useRemixForgeSocket";
+export {
+  initRouteBoundariesClient,
+  initRouteBoundariesServer,
+} from "./methods/boundaries";
