+++ src/index.ts
@@ -1,5 +1,15 @@
 import "./input.css";
-import { RemixDevTools, useRemixForgeSocket } from "./RemixDevTools";
-export { RemixDevTools, useRemixForgeSocket };
+import {
+  RemixDevTools,
+  useRemixForgeSocket,
+  initRouteBoundariesClient,
+  initRouteBoundariesServer,
+} from "./RemixDevTools";
+export {
+  RemixDevTools,
+  useRemixForgeSocket,
+  initRouteBoundariesClient,
+  initRouteBoundariesServer,
+};
 //export * from "./monitor";
 export default RemixDevTools;
