+++ src/remix-app-for-testing/app/root.tsx
@@ -86,7 +86,6 @@ export default function App() {
           <RemixDevTools
             additionalTabs={[plugin()]}
             defaultOpen
-            showRouteBoundaries
           />
         )}
       </body>
